requirements.txt
# Core requirements
Django>=5.0.0
pillow
python-dotenv
dj-database-url
django-allauth
djangorestframework
requests
django-cors-headers
whitenoise
gunicorn
oauthlib 
requests-oauthlib 
python3-openid 
PyJWT
cryptography
reportlab
six
psycopg2-binary
mysqlclient

# Authentication
django-allauth>=0.58.0,<0.60.0
oauthlib>=3.2.0,<4.0.0
requests-oauthlib>=1.3.0,<2.0.0
python3-openid>=3.2.0,<4.0.0
PyJWT>=2.8.0,<3.0.0

# REST API
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0

# OpenAI integration (optional)
openai>=1.0.0,<2.0.0
